Not working masking code with right frame freezed
This code was try to remove the ghosting by use of real time movement
tracking and then masking in overlapping region.
First frame is saved and all the next frames are used to calculate
difference which gives us the movement. Then two masks are created
Mask1> for right camera which will show the object in overlapping region
perfectly real time
Mask2> for left camera which will show the data from the saved first
frame
So this results in movement showing of only one camera(right) and static
scene of first frame from the other camera(left camera), hence removing
ghosting.